requirements_dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes

# Development requirements include dependencies to allow use of API
argparse==1.4.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
numpy==2.2.5
pandas==2.2.3
polygon-api-client==1.14.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==14.2
